requirements.txt
# Core dependencies for DSV Staff Map Generator
python-dotenv==1.2.1
pillow==11.3.0
opencv-python>=4.10.0
numpy>=1.26.0

# DSV Systems Integration
git+https://github.com/Edwinexd/dsv-wrapper.git

# Development tools
ruff>=0.7.0
pre-commit>=3.0.0
